fix: merged playground page into main page.
