+++ sites/static-image-api/package.json
@@ -21,11 +21,9 @@
 		"@types/request": "^2.48.12",
 		"@typescript-eslint/eslint-plugin": "^8.8.0",
 		"@typescript-eslint/parser": "^8.8.0",
-		"@undp-data/svelte-copy-to-clipboard": "^2.0.2",
-		"@undp-data/svelte-geohub-static-image-controls": "^2.0.21",
-		"@undp-data/svelte-sidebar": "^2.0.3",
-		"@undp-data/svelte-undp-components": "^1.3.1",
-		"@undp-data/svelte-undp-design": "^2.3.1",
+		"@undp-data/svelte-geohub-static-image-controls": "^2.0.24",
+		"@undp-data/svelte-undp-components": "^1.5.0",
+		"@undp-data/svelte-undp-design": "^2.3.2",
 		"@undp-data/undp-bulma": "^1.1.2",
 		"eslint": "^9.11.1",
 		"eslint-config-prettier": "^9.1.0",
@@ -50,7 +48,7 @@
 	"dependencies": {
 		"@mapbox/geo-viewport": "^0.5.0",
 		"@maplibre/maplibre-gl-native": "^6.0.0",
-		"pmtiles": "^3.2.0",
+		"pmtiles": "^3.2.1",
 		"sharp": "^0.33.5"
 	}
 }
