--- sites/static-image-api/src/routes/(app)/+page.ts
@@ -1,9 +1,0 @@
-import type { PageLoad } from './$types';
-
-export const load: PageLoad = async () => {
-	const title = 'GeoHub Static Image API';
-
-	return {
-		title
-	};
-};
