--- sites/static-image-api/src/routes/(app)/playground/+page.ts
@@ -1,9 +1,0 @@
-import type { PageLoad } from './$types';
-
-export const load: PageLoad = async () => {
-	const title = 'Playground | GeoHub Static API';
-
-	return {
-		title
-	};
-};
